Factor out callbacks from __init__.py.
